+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
   - Fix error when "Save As" to same location
+  - Fix error when cancelling "Save As"
 
 ## [2.2.4] - 2018-09-18
   - Fix loading error "processNextTick is not a function" for legacy .gko